requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
librosa==0.10.1
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
streamlit==1.28.0
soundfile==0.12.1
audioread==3.0.1
scipy==1.11.2
pillow==10.0.1
tqdm==4.66.1
joblib==1.3.2
python-dotenv==1.0.0
plotly==5.17.0
ipykernel==6.25.2
notebook==7.0.4
openpyxl==3.1.2